Add a tool that traverses and reports on a tree
This adds a TreeTraverser that can traverse a tree whose nodes have
values that are supported as arguments to the sum() function and returns
sums of these nodes' values for each level of the tree. The
TreeTraverser can also report the highest RowSum, including the row
number whose sum is highest and the total of that sum.